+++ lib/features/search/presentation/providers/geraete_suche_providers.g.dart
@@ -68,4 +68,4 @@ final class DurchsuchbarerBestandProvider
 }
 
 String _$durchsuchbarerBestandHash() =>
-    r'bcd37f0e00838c0f155ce167d7d670bf1b37f998';
+    r'b9b119c7d90bde340a89f6b71659d434eb5cda2a';